Move tallyVotes function to election utils.
