+++ packages/webapp/src/elections/utils.ts
@@ -1,5 +1,7 @@
 import dayjs from "dayjs";
 
+import { VoteData } from "elections/interfaces";
+
 export const extractElectionDates = (election: any) => {
     const rawStartDateTime = `${
         (election?.election_seeder && election.election_seeder.end_time) ||
@@ -22,5 +24,53 @@ export const extractElectionDates = (election: any) => {
         participationTimeLimit,
         estimatedEndDateTime,
         rawStartDateTime,
+    };
+};
+
+export const tallyVotesFromVoteData = (participantVoteData: VoteData[]) => {
+    const votes = participantVoteData.filter((pv) => pv.candidate);
+    const threshold = Math.floor(votes.length * (2 / 3) + 1);
+
+    const voteRecipients = new Set(
+        participantVoteData.map((v) => v.candidate).filter((c) => c)
+    );
+
+    const candidatesByVotesReceived: { [key: number]: string[] } = {};
+    Array.from(voteRecipients).forEach((candidate) => {
+        const numVotesReceived = participantVoteData.filter(
+            (vote) => vote.candidate === candidate
+        ).length;
+        if (candidatesByVotesReceived[numVotesReceived]) {
+            candidatesByVotesReceived[numVotesReceived].push(candidate);
+        } else {
+            candidatesByVotesReceived[numVotesReceived] = [candidate];
+        }
+    });
+
+    const leadTally = Object.keys(candidatesByVotesReceived).length
+        ? Math.max(...Object.keys(candidatesByVotesReceived).map(Number))
+        : 0;
+
+    const candidatesWithMostVotes = candidatesByVotesReceived[leadTally] ?? [];
+    const thereIsOneLeader = candidatesWithMostVotes.length === 1;
+
+    const leaderIsVotingForSelf =
+        thereIsOneLeader &&
+        participantVoteData.some(
+            (participant) =>
+                participant.member === candidatesWithMostVotes[0] &&
+                participant.candidate === candidatesWithMostVotes[0]
+        );
+
+    const didReachConsensus = leadTally >= threshold && leaderIsVotingForSelf;
+
+    return {
+        totalVotesCast: votes.length,
+        leadCandidates: candidatesWithMostVotes,
+        leadTally,
+        totalVotesRequiredForConsensus: threshold,
+        remainingVotesRequiredForConsensus: threshold - leadTally,
+        leaderIsVotingForSelf,
+        isThereConsensus: didReachConsensus,
     };
 };
